read server list from github

+++ src/com/boombuler/piraten/map/DatabaseHelper.java
@@ -8,7 +8,7 @@ import android.util.Log;
 public class DatabaseHelper extends SQLiteOpenHelper {
 
 	private static final String DATABASE_NAME = "plakate";
-	private static final int DATABASE_VERSION = 1;
+	private static final int DATABASE_VERSION = 2;
 	
 	
 	private static final String PLAKATE_CREATE = "create table "+DBAdapter.TABLE_PLAKATE+" (" +
@@ -22,7 +22,14 @@ public class DatabaseHelper extends SQLiteOpenHelper {
 	private static final String CHANGES_CREATE = "create table "+DBAdapter.TABLE_CHANGES + " (" +
 			DBAdapter.CHANGES_ID + " integer primary key, " +
 			DBAdapter.CHANGES_TYPE + " integer);";
-	
+    private static final String SERVERS_CREATE = "create table "+DBAdapter.TABLE_SERVERS + " (" +
+            DBAdapter.SERVERS_ID + " text primary key," +
+            DBAdapter.SERVERS_NAME + " text, " +
+            DBAdapter.SERVERS_INFO + " text, " +
+            DBAdapter.SERVERS_URL + " text, " +
+            DBAdapter.SERVERS_DEV + " integer not null" +
+            ");";
+
 	
 	public DatabaseHelper(Context context) {
 		super(context, DATABASE_NAME, null, DATABASE_VERSION);	
@@ -34,6 +41,7 @@ public class DatabaseHelper extends SQLiteOpenHelper {
 	public void onCreate(SQLiteDatabase database) {
 		database.execSQL(CHANGES_CREATE);
 		database.execSQL(PLAKATE_CREATE);
+        database.execSQL(SERVERS_CREATE);
 	}
 	
 	// Method is called during an upgrade of the database, e.g. if you increase
@@ -44,9 +52,15 @@ public class DatabaseHelper extends SQLiteOpenHelper {
 		Log.w(DatabaseHelper.class.getName(),
 				"Upgrading database from version " + oldVersion + " to "
 						+ newVersion + ", which will destroy all old data");
-		database.execSQL("DROP TABLE IF EXISTS "+DBAdapter.TABLE_PLAKATE);
-		database.execSQL("DROP TABLE IF EXISTS "+DBAdapter.TABLE_CHANGES);
-		onCreate(database);
+        if (oldVersion < 1) {
+		    database.execSQL("DROP TABLE IF EXISTS "+DBAdapter.TABLE_PLAKATE);
+		    database.execSQL("DROP TABLE IF EXISTS "+DBAdapter.TABLE_CHANGES);
+            database.execSQL(CHANGES_CREATE);
+            database.execSQL(PLAKATE_CREATE);
+        }
+        if (oldVersion < 2) {
+            database.execSQL(SERVERS_CREATE);
+        }
 	}
 
 }
